ath-reg:  Relax application of default world regd
When the driver specifies a particular country code, then
this code ends up using the default world regdb, which disables
a great deal of things.  CRDA never re-enables these channels,
I am not sure if it used to or not..the comments in this code indicated
it did I think.

So, just don't apply our regdb if a country code is specified, and let
the mac80211 stack take care of applying the regulatory settings based
on the country code and user-space crda.

Signed-off-by: Ben Greear <greearb@candelatech.com>