ath10k:  Support appending values to eeprom configAddr.
The general eeprom/board-data/cal flow is:  Load board.bin from disk,
set that in firmware, run OTP which will modify the
board.bin, boot up the main firmware which will use that
previously set eeprom data.

The problem is, if you want to have some special register config in
the board file, the OTP will probaby over-ride this, so it will not
actually apply to the firmware.

To work around this issue, save the extra config in the driver, and
then apply it to the firmware with a new ath10k-ct (wave-1) specific
API.

Signed-off-by: Ben Greear <greearb@candelatech.com>