ath10k:  Improve crash interaction with wmi timeout logic.
Need to set the crashed flag and also trigger the wmi wait logic to
complete so that logic can bail out quickly.  I think waiting on the work-queue
to do that will probably block on the wmi processing thread.  This caused us
to see wmi timeout logic kick in when fw crashed.

Signed-off-by: Ben Greear <greearb@candelatech.com>